requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers==4.38.2
pycocotools
scikit-learn
opencv-python
mlflow
tqdm
matplotlib
datasets
Pillow
timm